Add support for "local" programs
For my experimentation I am going to be writing programs which
are more than a single .c file, we stash those away in the
local directory, but otherwise follow the same convention (the
compiled program `_prog` will be installed as `prog`, the same
way user programs are.

Also fix mkfs entry in .gitignore.